+++ crates/core/tests/java/anonymous_class_bodies/chain_receiver.java
@@ -1,0 +1,9 @@
+class ChainReceiver {
+    void use() {
+        new Thread(new Runnable() {
+            public void run() {
+                go();
+            }
+        }).start();
+    }
+}