+++ crates/core/tests/java/anonymous_class_bodies/initialiser_control.out.java
@@ -1,0 +1,34 @@
+class InitialiserControl {
+    Runnable field = new Runnable() {
+        public void run() {
+            go();
+        }
+    };
+
+    void use() {
+        Runnable local = new Runnable() {
+            public void run() {
+                go();
+            }
+        };
+        new Runnable() {
+            public void run() {
+                go();
+            }
+        };
+        Runnable[] array = new Runnable[]{
+            new Runnable() {
+                public void run() {
+                    go();
+                }
+            }};
+    }
+
+    Runnable make() {
+        return new Runnable() {
+            public void run() {
+                go();
+            }
+        };
+    }
+}