+++ crates/core/tests/java/anonymous_class_bodies/ternary.out.java
@@ -1,0 +1,9 @@
+class Ternary {
+    Object use(boolean b) {
+        return b ? new Runnable() {
+            public void run() {
+                go();
+            }
+        } : null;
+    }
+}